pyproject.toml
[tool.poetry]
name = "assistant-core"
version = "0.0"
description = "The core of the LLM based assistant implemented in Python with langchain"
authors = ["Vicente Garofalo <vicente.garofalo@avangenio.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
tavily-python = "^0.5.0"
langchain = "^0.3.1"
langchain-core = "^0.3.6"
langgraph = "^0.2.15"
langchain-community = "^0.3.1"
langchain-fireworks = "^0.2.0"
langchain-openai = "^0.2.1"
openai = "^1.43.0"
python-dotenv = "^1.0.1"
mistralai = "^1.0.2"
langchain-mistralai = "^0.2.0"
pypdf = "^5.0.0"
ollama = "^0.3.2"
faiss-cpu = "^1.8.0.post1"
langchain-ollama = "^0.2.0"
mypy = "^1.11.2"
types-tqdm = "^4.66.0.20240417"
langchain-text-splitters = "^0.3.0"
langgraph-checkpoint = "^1.0.12"
langchain-huggingface = "^0.1.0"
types-pyyaml = "^6.0.12.20240917"
fastapi = "^0.115.0"
supabase = "^2.8.1"
uvicorn = "^0.31.0"
psycopg2 = "^2.9.9"
langchain-cohere = "^0.3.1"
rank-bm25 = "^0.2.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
isort = "^5.13.2"
black = "^24.8.0"
autoflake = "^2.3.1"
mypy = "^1.11.1"
types-passlib = "^1.7.7.20240819"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]
no_implicit_optional = true
warn_return_any = true
show_error_codes = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["storage3.*", "supabase"]
ignore_missing_imports = true
